requirements.txt
# Core Trading and Data Processing
pandas>=1.5.0
numpy>=1.21.0
python-binance>=1.0.16
ccxt>=4.0.0
TA-Lib>=0.4.25

# Machine Learning - Core
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0

# Machine Learning - Optional Advanced (install if available)
lightgbm>=3.3.0
catboost>=1.2.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0

# CUDA/GPU Acceleration (optional - install if NVIDIA GPU available)
# cudf-cu11>=23.0.0
# cuml-cu11>=23.0.0
# rapids-singlecell>=1.0.0

# Data Analysis and Statistics
scipy>=1.9.0
statsmodels>=0.14.0

# Database and Storage
sqlite3  # Built-in with Python
pickle5>=0.0.11; python_version < "3.8"

# Communication and Notifications
python-telegram-bot>=20.0
requests>=2.28.0
urllib3>=1.26.0

# Async and Threading
asyncio  # Built-in with Python
threading  # Built-in with Python
aiohttp>=3.8.0

# Environment and Configuration
python-dotenv>=1.0.0
pathlib  # Built-in with Python 3.4+

# Utilities
datetime  # Built-in with Python
time  # Built-in with Python
json  # Built-in with Python
hashlib  # Built-in with Python
uuid  # Built-in with Python
re  # Built-in with Python
os  # Built-in with Python
sys  # Built-in with Python
warnings  # Built-in with Python

# Data Visualization (optional)
matplotlib>=3.6.0
plotly>=5.15.0
seaborn>=0.12.0

# Additional ML Tools
imbalanced-learn>=0.10.0
optuna>=3.0.0  # For hyperparameter optimization

# Performance and Monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Jupyter Notebook Support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.5.0

# Additional Dependencies for Enhanced Features
tqdm>=4.64.0  # Progress bars
colorama>=0.4.6  # Colored terminal output
rich>=13.0.0  # Rich text and beautiful formatting

# Financial Data Analysis
yfinance>=0.2.0  # Alternative data source
alpha-vantage>=2.3.0  # Alternative data source

# Caching and Performance
redis>=4.5.0  # Optional Redis caching
diskcache>=5.6.0  # Disk-based caching alternative

# Configuration Management
pydantic>=2.0.0  # Data validation
configparser  # Built-in with Python

# Logging and Monitoring
loguru>=0.7.0  # Enhanced logging
structlog>=23.0.0  # Structured logging

# Time Series Analysis
arch>=5.3.0  # GARCH models
statsforecast>=1.4.0  # Time series forecasting

# Network and API
httpx>=0.24.0  # Modern HTTP client
websockets>=11.0.0  # WebSocket support for real-time data

# Data Validation and Cleaning
cerberus>=1.3.4  # Data validation
missingno>=0.5.2  # Missing data visualization

# Parallel Processing
concurrent.futures  # Built-in with Python 3.2+
multiprocessing  # Built-in with Python
